Fully MPI-enabled version
